requirements.txt
fastapi
uvicorn
requests
gspread
google-auth
openai
